# api_docs.yaml
openapi: 3.0.3
info:
  title: SafeBoda Rwanda API
  version: 1.0.0
  description: Comprehensive ride-booking platform API for Rwanda
paths:
...
  /api/bookings/{id}/status/:
    put:
      operationId: bookings_status_update
      description: Update ride status (driver arrival, start, completion, etc.)
      summary: Update booking status
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - Bookings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RideStatusUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RideStatusUpdateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RideStatusUpdateRequest'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RideStatusUpdate'
          description: ''
    patch:
      operationId: bookings_status_partial_update
      description: PUT /api/bookings/{id}/status/ - Update booking status
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - bookings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedRideStatusUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedRideStatusUpdateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedRideStatusUpdateRequest'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RideStatusUpdate'
          description: ''
...
components:
  schemas:
...
    PatchedRideStatusUpdateRequest:
      type: object
      description: Update ride status
      properties:
        status:
          $ref: '#/components/schemas/Status493Enum'
        reason:
          type: string
          writeOnly: true
          minLength: 1
...
    RideStatusUpdate:
      type: object
      description: Update ride status
      properties:
        status:
          $ref: '#/components/schemas/Status493Enum'
    RideStatusUpdateRequest:
      type: object
      description: Update ride status
      properties:
        status:
          $ref: '#/components/schemas/Status493Enum'
        reason:
          type: string
          writeOnly: true
          minLength: 1
...
    Status493Enum:
      enum:
      - requested
      - driver_assigned
      - driver_arrived
      - in_progress
      - completed
      - cancelled_by_customer
      - cancelled_by_driver
      - cancelled_by_system
      - no_driver_found
      type: string
      description: |-
        * `requested` - Ride Requested
        * `driver_assigned` - Driver Assigned
        * `driver_arrived` - Driver Arrived
        * `in_progress` - Ride in Progress
        * `completed` - Ride Completed
        * `cancelled_by_customer` - Cancelled by Customer
        * `cancelled_by_driver` - Cancelled by Driver
        * `cancelled_by_system` - Cancelled by System
        * `no_driver_found` - No Driver Found
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT